+++ server/data-loader/index.ts
@@ -2,7 +2,10 @@ import {createConnection, getConnectionOptions, getRepository} from "typeorm";
 
 import {PlayerStat} from '../domain';
 
+import * as data from  '../rushing.json';
+
 async function loadData(){
+    console.log("Loading data from file");
     let connectionOptions = await getConnectionOptions();
 
     connectionOptions = {
@@ -13,9 +16,21 @@ async function loadData(){
     }
 
     await createConnection(connectionOptions);
-    
+
+    const statRepository = getRepository(PlayerStat); 
+    const cleanData = data.map(d => {
+        try{
+            return PlayerStat.fromJson(d)
+        } catch(e){
+            console.log('ERRO')
+            console.log(d)
+            console.log(e)
+            throw e;
+        }
+    });
+    await statRepository.save(cleanData)
+    console.log("Finished loading the data")
 }
 
 loadData()
-    .then(() => console.log("Finished load the data"))
     .catch((e) => console.log(e))